Refactor: Removed mockAPI
